requirements.txt
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.5.0
pytest-timeout==2.2.0
allure-pytest==2.13.2
pandas==2.2.0
openpyxl==3.1.2
pyyaml==6.0.1
requests==2.31.0
colorlog==6.8.0
